requirements.txt
aiogram==2.20
aioredis==2.0.1
asyncpg==0.25.0
python-decouple==3.6
psycopg2-binary==2.9.3
python-decouple==3.6
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
dj-database-url==0.5.0
gunicorn==20.1.0